% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein.R
\name{kmat}
\alias{kmat}
\title{Compute a Stein kernel matrix}
\usage{
kmat(x, s, vfk0)
}
\arguments{
\item{x}{n-by-d matrix of n d-dimensional sample points.}

\item{s}{n-by-d matrix of gradients of log target at \code{x}.}

\item{vfk0}{vectorised Stein kernel function.}
}
\value{
n-by-n Stein kernel matrix.
}
\description{
Compute a Stein kernel matrix
}
