% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein.R
\name{ksd}
\alias{ksd}
\title{Compute a cumulative sequence of KSD values}
\usage{
ksd(x, s, vfk0)
}
\arguments{
\item{x}{n-by-d matrix of n d-dimensional sample points.}

\item{s}{n-by-d matrix of gradients of log target at \code{x}.}

\item{vfk0}{vectorised Stein kernel function.}
}
\value{
n-vector of cumulative KSD values.
}
\description{
Compute a cumulative sequence of KSD values
}
