% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{make_imq}
\alias{make_imq}
\title{Create the (IMQ) Stein kernel function with a preconditioner}
\usage{
make_imq(smp, scr, pre = "sclmed")
}
\arguments{
\item{smp}{n-by-d matrix of n d-dimensional sample points.}

\item{scr}{n-by-d matrix of gradients of log target at \code{smp}.}

\item{pre}{string specifying the heuristic for computing the
preconditioner matrix, either 'med', 'sclmed', or 'smpcov'.
Alternatively, a numeric string can be passed as the single
length-scale parameter of an isotropic kernel.}
}
\value{
Stein kernel function.
}
\description{
Create the (IMQ) Stein kernel function with a preconditioner
}
