% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thin}
\alias{thin}
\title{Perform Stein thinning}
\usage{
thin(smp, scr, m, pre = "sclmed")
}
\arguments{
\item{smp}{n-by-d matrix of n d-dimensional sample points.}

\item{scr}{n-by-d matrix of gradients of log target at \code{smp}.}

\item{pre}{string specifying the heuristic for computing the
preconditioner matrix, either 'med', 'sclmed', or 'smpcov'.
Alternatively, a numeric string can be passed as the single
length-scale parameter of an isotropic kernel.}
}
\value{
m-vector containing the row indices in \code{smp} (and \code{scr})
of the selected points.
}
\description{
Optimally select m points from n > m samples generated from a target
distribution of d dimensions.
}
