% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{vfk0_imq}
\alias{vfk0_imq}
\title{Stein kernel function}
\usage{
vfk0_imq(a, b, sa, sb, linv)
}
\arguments{
\item{a}{first argument of the kernel, m-by-d matrix.}

\item{b}{second argument of the kernel, m-by-d matrix.}

\item{sa}{gradients of the log target density at \code{a}, m-by-d matrix.}

\item{sb}{gradients of the log target density at \code{b}, m-by-d matrix.}

\item{linv}{inverse preconditioner, d-by-d matrix.}
}
\value{
value of the Stein kernel, m-vector.
}
\description{
Vectorised Stein kernel function based on the inverse multi-quadric (IMQ)
kernel. Each argument can be m-by-d matrices containing m d-dimensional
points.
}
