% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{make_imq}
\alias{make_imq}
\title{Create the (IMQ) Stein kernel function with a preconditioner}
\usage{
make_imq(smp, scr, pre = "id")
}
\arguments{
\item{smp}{n-by-d matrix of n d-dimensional sample points.}

\item{scr}{n-by-d matrix of gradients of log target at \code{smp}.}

\item{pre}{optional string, either "id" (default), "med", "sclmed", or
"smpcov", specifying the preconditioner to be used. Alternatively,
a numeric string can be passed as the single length-scale parameter
of an isotropic kernel.}
}
\value{
Stein kernel function.
}
\description{
Create the (IMQ) Stein kernel function with a preconditioner
}
