% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thin}
\alias{thin}
\title{Perform Stein thinning}
\usage{
thin(smp, scr, m, stnd = TRUE, pre = "id")
}
\arguments{
\item{smp}{n-by-d matrix of n d-dimensional sample points.}

\item{scr}{n-by-d matrix of gradients of log target at \code{smp}.}

\item{stnd}{optional logical, either TRUE (default) or FALSE, indicating
whether or not to standardise the colums of \code{smp}.}

\item{pre}{optional string, either "id" (default), "med", "sclmed", or
"smpcov", specifying the preconditioner to be used. Alternatively,
a numeric string can be passed as the single length-scale parameter
of an isotropic kernel.}
}
\value{
m-vector containing the row indices in \code{smp} (and \code{scr})
of the selected points.
}
\description{
Optimally select m points from n > m samples generated from a target
distribution of d dimensions.
}
